server/requirements.txt
tensorflow==2.14.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
joblib==1.3.2
matplotlib
numpy